backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv>=1.0.1
motor==3.3.1
pydantic>=2.6.4
pymongo==4.5.0
certifi>=2023.7.22
pyopenssl>=23.2.0
